games/spoiled_broth/requirements.txt
eventlet==0.40.2
Flask==3.1.1
numpy==2.3.1
Pillow==11.3.0
Werkzeug==3.1.3
gunicorn==21.2.0
gevent==23.9.1
flask-socketio==5.3.6
protobuf==4.23.4
google-cloud-storage==3.2.0
orjson==3.10.7
redis
